+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 ## [unreleased]
  ### Added 
  ### Fixed 
+* Deprecate the outdated QueryDefinitionListResponseData class ([#661](https://github.com/ehrbase/openEHR_SDK/pull/661))
 
 ## [2.20.0]
  ### Changed